pyproject.toml
[project]
name = "HACX"
version = "0.1.0"
description = "ViT-B/16 training on HacX dataset"
authors = [
    {name = "Aloysius Chia", email = "aloysiuschia495@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0", 
    "transformers>=4.30.0",
    "timm>=0.9.0",
    "pillow>=9.0.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "tqdm>=4.64.0",
    "wandb>=0.15.0",
    "accelerate>=0.20.0",
    "evaluate>=0.4.0",
"pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0", 
    "pytest>=7.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]